Added an algorithm to remove the current Row.
(In the previous commit are located these changes)

I had to add an data attribute to every Row that is created, and make it the same number of his position inside the Library array.
Then, i created a function called deleteBook, which takes the number of the last element inside of the Library array and then i "Queryselected" the Row that we want to errase by using his data attribute number.
With the row selected, now we can erase it using remove()