Fixed the delete button, deleting the last row.
Before, the deleteBook function always deleted the last row, because i was calculating the currentPosition with the last book of myLibrary.
So in order to fix this, i assigned the same dataId to every row and his delete button when creating them.
Now we can get the specific dataId of the delete button clicked event with e.target.dataset.id, with this information, we know which row we have to erase.