Added a toggleRead function to change read status.
First, it simply changes the text that the user can read in screen, alternating between Yes or No.
Then, i needed to know the data id of the specific button that is being clicked, so i can call the changeRead() "property" function in the Book i need in specific.
What changeRead() does is that it changes the item3 value which is the read value inside of the actual book object, so when we change it on screen, it actually changes too inside the Book object.